requirements.txt
Flask
Flask-Cors
PyJWT
cloud-sql-python-connector
SQLAlchemy
requests
python-dotenv
PyMySQL
gunicorn